se agrega propiedades sobre el objeto jsonObject y se valida existencia de archivo

+++ sesion6/Archivo.ts
@@ -4,8 +4,26 @@ import fs from "fs";
 export class Archivo {
 
     private filename: string;
-    private jsonObject: any = [];
+    private _jsonObject: any = [];
 
+    /**
+     * getter jsonObject
+     */
+    get jsonObject():[] {
+        return this._jsonObject;
+    }
+
+    /**
+     * setter jsonObject
+     */
+    set jsonObject(value:[]){
+        this._jsonObject = value;
+    }
+
+    /**
+     * 
+     * @param filename 
+     */
     constructor(filename: string) {
         this.filename = filename;
     }
@@ -16,9 +34,9 @@ export class Archivo {
     public async leer() {
         try {
             const contenido = await fs.promises.readFile(this.filename, 'utf-8');
-            this.jsonObject = JSON.parse(contenido);
+            this._jsonObject = JSON.parse(contenido);
             console.log(this.jsonObject);
-            
+
         } catch (error) {
             console.error(error);
         }
@@ -28,6 +46,7 @@ export class Archivo {
      * guardar un nuevo objeto usando la libreria chance que retorna elementos random
      */
     public async guardar() {
+        let isExist = false;
         const chance = new Chance();
         const newElement = {
             title: chance.animal(),
@@ -35,17 +54,36 @@ export class Archivo {
             thumbmail: chance.avatar(),
             id: this.jsonObject.length + 1
         }
-        this.jsonObject.push(newElement);
+        this._jsonObject.push(newElement);
 
         try {
-            await fs.promises.writeFile(this.filename,
-                JSON.stringify(this.jsonObject, null, '\t'), 'utf-8');
+            isExist = await this.checkIfExist(this.filename);
+            if (!isExist) {
+                await fs.promises.writeFile(this.filename,
+                    JSON.stringify(this.jsonObject, null, '\t'), 'utf-8');
+            } else {
+                await fs.promises.appendFile(this.filename,
+                    JSON.stringify(this.jsonObject, null, '\t'), 'utf-8');
+            }
         } catch (error) {
             console.error(error);
         }
 
     }
 
+    /**
+     * 
+     * @param filename 
+     * @returns true si existe el archivo
+     */
+    private async checkIfExist(filename: string) {
+        try {
+            await fs.promises.access(this.filename);
+        } catch (error) {
+            return false;
+        }
+        return true;
+    }
 
     /**
      * borrar
